+++ Chauffeur.Tests/Deliverables/UpgradeDeliverableTests.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Linq;
+using System.Xml;
 using Chauffeur.Deliverables;
+using Chauffeur.Host;
 using Chauffeur.Services;
 using NSubstitute;
 using Semver;
@@ -15,13 +17,23 @@ namespace Chauffeur.Tests.Deliverables
 {
     public class UpgradeDeliverableTests
     {
+        private readonly MockTextWriter writer;
+        private readonly IChauffeurSettings settings;
+        private readonly IXmlDocumentWrapper xmlDocumentWrapper;
+        private readonly IMigrationRunnerService migrationRunnerService;
+        
+        public UpgradeDeliverableTests()
+        {
+            writer = new MockTextWriter();
+            settings = Substitute.For<IChauffeurSettings>();
+            xmlDocumentWrapper = Substitute.For<IXmlDocumentWrapper>();
+            migrationRunnerService = Substitute.For<IMigrationRunnerService>();
+        }
+
         [Fact]
         public async Task NoMigrationEntries_WillWarnAndExit()
         {
-            var writer = new MockTextWriter();
-            
-            var deliverable = new UpgradeDeliverable(null, writer, null, MigrationEntryService());
-
+            var deliverable = new UpgradeDeliverable(null, writer, null, MigrationEntryService(), settings, xmlDocumentWrapper);
             var response = await deliverable.Run(null, null);
 
             Assert.Single(writer.Messages);
@@ -31,10 +43,8 @@ namespace Chauffeur.Tests.Deliverables
         [Fact]
         public async Task SameVersions_NothingDone()
         {
-            var writer = new MockTextWriter();
-            
-            var deliverable = new UpgradeDeliverable(null, writer, null, MigrationEntryService(UmbracoVersion.GetSemanticVersion()));
 
+            var deliverable = new UpgradeDeliverable(null, writer, null, MigrationEntryService(UmbracoVersion.GetSemanticVersion()), settings, xmlDocumentWrapper);
             var response = await deliverable.Run(null, null);
 
             Assert.Single(writer.Messages);
@@ -44,9 +54,9 @@ namespace Chauffeur.Tests.Deliverables
         [Fact]
         public async Task CorrectSameVersionSelected_NothingDone()
         {
-            var writer = new MockTextWriter();
-            
-            var deliverable = new UpgradeDeliverable(null, writer, null, MigrationEntryService(new SemVersion(7, 1), UmbracoVersion.GetSemanticVersion(), new SemVersion(7,3)));
+
+            var deliverable = new UpgradeDeliverable(null, writer, null, 
+                MigrationEntryService(new SemVersion(7, 1), UmbracoVersion.GetSemanticVersion(), new SemVersion(7,3)), settings, xmlDocumentWrapper);
 
             var response = await deliverable.Run(null, null);
 
@@ -57,12 +67,27 @@ namespace Chauffeur.Tests.Deliverables
         [Fact]
         public async Task DifferentVersions_UpgradeIsCompleted()
         {
-            var writer = new MockTextWriter();
-
-            var migrationRunnerService = Substitute.For<IMigrationRunnerService>();
             migrationRunnerService.Execute(Arg.Any<SemVersion>(), Arg.Any<SemVersion>()).Returns(true);
+            var anyStringArg = Arg.Any<string>();
+            settings.TryGetSiteRootDirectory(out anyStringArg).Returns( 
+                x => {
+                    x[0] = @"C:\test\umbraco";
+                    return true;
+                });
 
-            var deliverable = new UpgradeDeliverable(null, writer, migrationRunnerService, MigrationEntryService(new SemVersion(7,1)));
+            var xmlDoc = new XmlDocument();
+            xmlDoc.LoadXml(
+                 "<configuration>" +
+                 "<appSettings>" +
+                 "<add key=\"umbracoConfigurationStatus\" value=\"1.1.1\" />" +
+                 "</appSettings>" +
+                 "</configuration>"
+                 );
+
+            xmlDocumentWrapper.LoadDocument(Arg.Any<string>()).Returns(xmlDoc);
+
+            var deliverable = new UpgradeDeliverable(null, writer, migrationRunnerService, 
+                MigrationEntryService(new SemVersion(7,1)), settings, xmlDocumentWrapper);
 
             var response = await deliverable.Run(null, null);
 
@@ -73,10 +98,10 @@ namespace Chauffeur.Tests.Deliverables
         [Fact]
         public async Task UpgradeFailes_ErrorStateReturned()
         {
-            var writer = new MockTextWriter();
-            var migrationRunnerService = Substitute.For<IMigrationRunnerService>();
             migrationRunnerService.Execute(Arg.Any<SemVersion>(), Arg.Any<SemVersion>()).Returns(false);
-            var deliverable = new UpgradeDeliverable(null, writer, migrationRunnerService, MigrationEntryService(new SemVersion(7, 1)));
+
+            var deliverable = new UpgradeDeliverable(null, writer, migrationRunnerService, 
+                MigrationEntryService(new SemVersion(7, 1)), settings, xmlDocumentWrapper);
 
             var response = await deliverable.Run(null, null);